Move display related functions to tools
Some functions in the dfc are not directly related to dfc but more general, so I moved them to tools
