+++ .gitignore
@@ -35,3 +35,6 @@ stmpy/.DS_Store
 stmpy/doc/.DS_Store
 stmpy/.ipynb_checkpoints/tools-checkpoint.py
 stmpy/.ipynb_checkpoints/driftcorr-checkpoint.py
+stmpy/.DS_Store
+stmpy/.ipynb_checkpoints/tools_old-checkpoint.py
+stmpy/doc/.DS_Store
